Massive overhaul of memory management.
This commit removes the mmu class in favor of baking the addressing
functionality in with the cpu. As a result, Ppu now owns its VRAM
vector, and all memory-mapped I/O now belongs to the "hardware" that the
memory is mapped to. For example, the actual registers for DIV, TIMA,
TMA, and TAC are found in the Timer class, and the Cpu will
appropriately query the Timer class for that value (as opposed to
everything being opaque bytes in Mmu's memory_ vector). Additionally,
work RAM, et. al., have been moved to the new, simplified Memory class.

Ideally, this change helps massage the dependency tree into a nice
place. So far, it's been tested successfully by the Blargg ROMs.

Additionally, stack-based storage is replaced with heap-based storage.